requirements.txt
flask-multipass[saml]>=0.4.*
indico>=3.2
flask_sqlalchemy